refactor: 웨이팅 정보 패키지 구조 변경, 화면 전환 앵니메이션, 각 navigation으로 이동
* **refactor: 웨이팅 정보(WaitingInfo) 관련 파일을 `setting` 패키지 하위로 이동**
    * 웨이팅 정보 관리 기능이 설정 도메인에 종속됨에 따라 관련 파일들의 패키지 경로를 수정했습니다.
    * `presentation.waitinginfo` -> `presentation.setting.waitinginfo`
    * `WaitingInfoViewModel`, `WaitingInfoUiState`, `AddWaitingInfoScreen` 등의 위치를 변경하고 관련 임포트 문을 업데이트했습니다.

* **style: 메인 탭 네비게이션 전환 애니메이션 제거**
    * `MainScreen`의 전역 설정으로 되어 있던 전환 애니메이션을 제거하고, 각 개별 탭(`Home`, `News`, `Schedule`, `PlaceMap`, `Setting`)의 `composable` 설정에 `EnterTransition.None` 및 `ExitTransition.None`을 명시적으로 적용했습니다.
    * 탭 전환 시 시각적 지연 없이 즉각적으로 화면이 전환되도록 개선했습니다.

* **chore: 코드 정리 및 경로 최적화**
    * `MainScreen.kt`, `SettingNavigation.kt` 등에서 변경된 패키지 경로에 맞춰 의존성 참조를 수정했습니다.
