+++ composeApp/src/commonMain/kotlin/com/daedan/festabook/presentation/setting/navigation/SettingNavigation.kt
@@ -1,5 +1,7 @@
 package com.daedan.festabook.presentation.setting.navigation
 
+import androidx.compose.animation.EnterTransition
+import androidx.compose.animation.ExitTransition
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.ui.Modifier
@@ -11,8 +13,8 @@ import com.daedan.festabook.presentation.main.FestabookRoute
 import com.daedan.festabook.presentation.main.MainTabRoute
 import com.daedan.festabook.presentation.setting.SettingViewModel
 import com.daedan.festabook.presentation.setting.component.SettingRoute
-import com.daedan.festabook.presentation.waitinginfo.WaitingInfoViewModel
-import com.daedan.festabook.presentation.waitinginfo.component.AddWaitingInfoRoute
+import com.daedan.festabook.presentation.setting.waitinginfo.WaitingInfoViewModel
+import com.daedan.festabook.presentation.setting.waitinginfo.component.AddWaitingInfoRoute
 
 fun NavGraphBuilder.settingNavGraph(
     innerPadding: PaddingValues,
@@ -25,7 +27,10 @@ fun NavGraphBuilder.settingNavGraph(
     onNavigateToAddWaitingInfo: () -> Unit,
     onBackClick: () -> Unit,
 ) {
-    composable<MainTabRoute.Setting> {
+    composable<MainTabRoute.Setting>(
+        enterTransition = { EnterTransition.None },
+        exitTransition = { ExitTransition.None },
+    ) {
         SettingRoute(
             modifier = Modifier.padding(innerPadding),
             homeViewModel = homeViewModel,
